Default

+++ onmt/Loss.py
@@ -246,8 +246,8 @@ class NMTLossCompute(LossComputeBase):
         loss_data_unk = loss_unk.data.clone()
         loss_data = loss.data.clone()
         LOSS =loss+loss_unk
-        target_unk =target_unk * Variable(target_unk.data.ne(self.unk_idx).float())
-        target = target * Variable(target_unk.data.eq(self.unk_idx).float())
+        target_unk =target_unk * Variable(target_unk.data.ne(self.unk_idx).long())
+        target = target * Variable(target_unk.data.eq(self.unk_idx).long())
         stats = self._stats(loss_data_unk, loss_data, scores_unk.data, scores.data, target_unk.view(-1).data, target.view(-1).data)
 
         return LOSS, stats
